demos/fraud-detection/requirements.txt
glassflow>=3.6.0
python-dotenv
